requirements.txt

streamlit>=1.36
python-docx>=1.1.0
docxcompose>=1.4.0
PyYAML>=6.0.1
